Reveal App/backend/requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.0.5
django-cors-headers==3.2.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
entrypoints==0.3
flake8==3.7.9
gunicorn==20.0.4
idna==2.9
itypes==1.1.0
Jinja2==2.11.2
Markdown==3.2.1
MarkupSafe==1.1.1
mccabe==0.6.1
pdfminer==20191125
Pillow==7.1.1
psycopg2==2.8.5
pycodestyle==2.5.0
pycryptodome==3.9.7
pyflakes==2.1.1
PyJWT==1.7.1
pytz==2019.3
requests==2.23.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.8
